+++ src/bracket-jumper.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { Utilities } from './lib';
+import { Utilities } from './metago.lib';
 
 class bracket {
     constructor(public start: string, public end: string) { }