+++ src/metago.lib/index.ts
@@ -1,0 +1,5 @@
+export * from './color';
+export * from './cursor';
+export * from './utilities';
+export * from './viewport'
+export * from './command-indicator'